Migrate Mutation.createMemberRole
